create user controller and register a user

+++ src/utils/cloudinary.js
@@ -2,27 +2,25 @@ import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
 cloudinary.config({
-  cloud_name: process.env.CLOUDINARY_NAME,
-  api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_API_SECRET,
+  cloud_name: "dcs3oqrqn",
+  api_key: "132727459123953",
+  api_secret: "e3vE9Bk72h19WfnWfrP83XTwmJA",
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
   try {
     if (!localFilePath) return null;
-    // upload file on the cloudinary
 
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
     });
 
-    // file has been uploaded successfully
-    console.log("file upload succesfully", response.url);
-
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the localy save temporary file as the upload operation got field
-
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+    console.error("Cloudinary upload error:", error.message);
     return null;
   }
 };
